feat: gate cache-edit microcompact and implement protocol source lifecycle events
* Gate cache-edit microcompact based on provider profile
* Implement Injected/Unregistered events for protocol skills and hooks
* Split thread rollout and turn logs into rara-persistence crate
* Update TUI context display with observability section
