+++ crates/rara-persistence/src/lib.rs
@@ -3,3 +3,5 @@ pub mod file_lock;
 pub mod redaction;
 pub mod thread_data;
 pub mod thread_metadata;
+pub mod thread_rollout_log;
+pub mod thread_turn_log;
