+++ src/hook_registry.rs
@@ -64,15 +64,13 @@ impl HookRegistry {
             }
             HookControlRequest::QueryHooks => {
                 let hooks = self.hooks.read().await;
-                for (hook_id, _entry) in hooks.iter() {
-                    // Emit a Declared event per hook so clients can reconstruct
-                    // the full list. There is no bulk-list variant in HookEvent.
-                    let _ =
-                        self.event_bus
-                            .publish_control(RuntimeEvent::Hook(HookEvent::Declared {
-                                hook_id: hook_id.clone(),
-                                lifecycle: HookLifecycle::SessionStart,
-                            }));
+                for (hook_id, entry) in hooks.iter() {
+                    let _ = self.event_bus.publish_control(RuntimeEvent::Hook(
+                        HookEvent::Declared {
+                            hook_id: hook_id.clone(),
+                            lifecycle: entry.lifecycle.clone(),
+                        },
+                    ));
                 }
             }
         }
