+++ src/runtime_control.rs
@@ -476,6 +476,8 @@ pub enum PromptSourceEvent {
 #[serde(tag = "type", content = "payload", rename_all = "snake_case")]
 pub enum SkillEvent {
     Registered { source_id: String, name: String },
+    Unregistered { source_id: String, name: String },
+    Injected { source_id: String, name: String },
     Shadowed { name: String, by_source_id: String },
     Failed { source_id: String, reason: String },
 }
@@ -560,6 +562,9 @@ pub enum HookEvent {
     Declared {
         hook_id: String,
         lifecycle: HookLifecycle,
+    },
+    Injected {
+        hook_id: String,
     },
     Ignored {
         hook_id: String,
@@ -1088,6 +1093,8 @@ mod tests {
                                 "enabled": true,
                                 "budget_chars": 48000,
                                 "keep_recent": 6,
+                                "cache_edit_eligible": false,
+                                "cache_edit_applied": false,
                                 "original_chars": 60000,
                                 "projected_chars": 30000,
                                 "saved_chars": 30000,
