+++ src/thread_store.rs
@@ -5,16 +5,16 @@ use std::path::{Path, PathBuf};
 
 use anyhow::{Context, Result, anyhow};
 use rara_persistence::atomic_file;
-use rara_persistence::thread_metadata;
-use rara_state::state_db::{
+use rara_persistence::thread_data::{
     PersistedCompactState, PersistedInteraction, PersistedLegacyRolloutMigration,
     PersistedLegacyRolloutSource, PersistedPlanStep, PersistedPromptRuntimeState,
     PersistedRecentThreadRecord, PersistedRuntimeRolloutItem, PersistedStructuredRolloutEvent,
     PersistedThreadLineage, PersistedThreadRecord, PersistedTurnEntry, PersistedTurnSummary,
-    StateDb,
 };
-use rara_state::thread_rollout_log::{self, RolloutEventRecorder};
-use rara_state::thread_turn_log;
+use rara_persistence::thread_metadata;
+use rara_persistence::thread_rollout_log::{self, RolloutEventRecorder};
+use rara_persistence::thread_turn_log;
+use rara_state::state_db::StateDb;
 use uuid::Uuid;
 
 use crate::agent::Message;
